Starting on Group summary list
Postgres TEXT type isn't automatically mapping for a projection, which
isn't immediately apparent from the misleading error messages and
lack of easily findable documentation about the projections/postgres
combination (or even projections) in general.

Aaaaanyway, forcing the selected type to varchar fixes the problem
because varchar is a standard SQL type and so the mapper handles it
properly.