Include unmatched OASys mappings in the table.
For OASys Fixed fields, we still need to supply a mapping code. The
difference is we provide the codes, rather than OASys providing them.
(For these fields, there's a mapping layer in the OASys API).

So, if we have a code in the source CSV that doesn't match a known OASYS
dynamica question, stick the code in the table anyway.