Add AssessmentEpisodeDto.errors field
If updating an episode results in any validation errors, whether
generated within ARN or received from OASys, the resulting
AssessmentEpisodeDto's errors field will be populated. The field maps
question UUIDs to an array of error messages for that question.

When unpopulated it is null rather than empty so that it doesn't get
serialised to JSON when passed up through the controllers.