Map oasys errors into the AssessmentEpisodeDto
When the AssessmentService gets any validation errors back from OASys,
it maps them back to the ARN questions and they get passed back up into
the front end.

To help with this mapping, QuestionService.getAllQuestions() now
returns QuestionSchemaEntities, rather than a naked list. This is a
little shim class that helps with mapping questions to OASys refs and
back.

This is a bit of crunchy commit. Sorry.