Arn 340 passing validation errors up to the front-end from OASys (#110)
* Add AssessmentEpisodeDto.errors field

If updating an episode results in any validation errors, whether
generated within ARN or received from OASys, the resulting
AssessmentEpisodeDto's errors field will be populated. The field maps
question UUIDs to an array of error messages for that question.

When unpopulated it is null rather than empty so that it doesn't get
serialised to JSON when passed up through the controllers.

* Pass the oasys update result back from updateOASysAssessment

How else are we going to know if it worked?

* Map oasys errors into the AssessmentEpisodeDto

When the AssessmentService gets any validation errors back from OASys,
it maps them back to the ARN questions and they get passed back up into
the front end.

To help with this mapping, QuestionService.getAllQuestions() now
returns QuestionSchemaEntities, rather than a naked list. This is a
little shim class that helps with mapping questions to OASys refs and
back.

* We should probably save our updated episode to the database

* Episode update return 200 or 422 if there are validation errors

This is a bit of crunchy commit. Sorry.