Support 'no input' as an input type (!). Trim question text.
'no input' seems paradoxical for an input type, but we do have question
text to which business rules are attached, but which have no input of
thier own. THey're not quite headings, and they're not quite information
text. Because the business rules are attached, and consequently they may
have error messages attached, I've chosen to treat that as questions
with no input.

They'll be corresponding change in hmpps-risk-assessment-ui to support
this.