Arn 370 ROSH Previous Convictions Questions (#136)
* Support 'no input' as an input type (!). Trim question text.

'no input' seems paradoxical for an input type, but we do have question
text to which business rules are attached, but which have no input of
thier own. THey're not quite headings, and they're not quite information
text. Because the business rules are attached, and consequently they may
have error messages attached, I've chosen to treat that as questions
with no input.

They'll be corresponding change in hmpps-risk-assessment-ui to support
this.

* Convert question 129.1 to no input

'Has the individual been convicted of any of the folowing ..."

* Expand the Murder convictions Q into current and previous.

* Expand out more of the risk to others ROSH questions