Pass MappingProvider into OasysAnswers.from
Rather than passing the list of all the questions pass the
MappingProvider, which we can then query for the questions. Why? Well,
in the next commit or two, we're also going to need to query which
questions are in a table.

Why MappingProvider rather than QuestionService? Focus and coupling -
OasysAnswers only needs a small subset of what QuestionService does,