Shape tables into the OasysAnswers
Tables need to be handled separately from the rest of the answers -
they're like a little horizontal chunk in an otherwise vertical list of
key-value pairs.

First, we identify which tables might be in play for this whole set of
answers. Table-by-table, we grab the relevant questions and associated
answers, and shape into OasysAnswers. That done, we process the
remaining answers as we did before.