Arn 395 update children in oasys (#150)
* Split AssessmentServiceTest class up into small chunks

* clean up OASys update tests a bit

better names
helper functions for set up
put mock set up in test make more explicit
blank lines to try and emphasis arrange/act/asset more clearly

* Exercise OasysAnswers.from directly

Was called from wider context, and inner function also tested, but not
by itself.

* Test for ARN->OASys with tabular answers

* Pass MappingProvider into OasysAnswers.from

Rather than passing the list of all the questions pass the
MappingProvider, which we can then query for the questions. Why? Well,
in the next commit or two, we're also going to need to query which
questions are in a table.

Why MappingProvider rather than QuestionService? Focus and coupling -
OasysAnswers only needs a small subset of what QuestionService does,

* Shape tables into the OasysAnswers

Tables need to be handled separately from the rest of the answers -
they're like a little horizontal chunk in an otherwise vertical list of
key-value pairs.

First, we identify which tables might be in play for this whole set of
answers. Table-by-table, we grab the relevant questions and associated
answers, and shape into OasysAnswers. That done, we process the
remaining answers as we did before.

* Remove duplicated setup for tests with children

* Overhauled OasysAnswers