Rework AnswerEntity from Collection of string to Collection of Answer
In this first step, the Answer object is a single string value, so
there's only a relatively small change to the code, and no observable
change in behaviour.

Next steps are to make the Answer itself a collection of Strings, so an
AnswerEntity could contain, for example, ["Animal", "Vegetable",
"Mineral", "Ghost"] or ["Dog"] or "["Trousers", ["Socks", "Shoes"], "Top"]

This last case is what we need to be able to properly support
multi-value answers in table rows.