ARN 445 - table row multi value answers (#168)
Converting AnswerEntity from containing a collection of strings to containing a collection of Answer, where an Answer is itself a collection of strings.

AnswerEntity could contain, for example, ["Animal", "Vegetable", "Mineral", "Ghost"] or ["Dog"] or ["Trousers", ["Socks", Shoes"], "Top"].

This last case is what we need to be able to properly support multi-value answers in table rows. A table row has a number of questions, eg child name, child age, pets. The answers are stored as

{
  child name: [["Bill"], ["Charlie"], ["Frank"]],
  child age: [["11"], ["13"], ["15"],
  pets: [["Binky mouse"], ["Goldie the goldfish", "Hammy the hamster"], ["Rambo"]]
}

Charlie has two pets, and we need to keep them together in the array of answers.

As part of this change, I've separated out AnswerEntity from (the new) AnswerDto. Previously, AnswerEntity was being pressed into two slightly different roles - serialising to the database, and serialising to JSON through the Controller. As it turns out, we need slightly different behaviour in these two areas. In particularly, we need AnswerEntity to maintain the structure of the Answers it contains - some of this is done by removing a number of constructors in favour of factory methods. The behaviour is there for us to use, but it stops the deserialisation using it.

A slightly curious asymmetry in our JSON handling is that incoming answers don't have the complex shape described above. The incoming answers for the second child in the above example would look like

{
  child name: "Charlie",
  child age: "13", 
  pets: ["Goldie the goldfish", "Hammy the hamster"]
}

We only ever add or update a single row at a time, so we can handle this these new answers are added into the table answers.